+++ src/main/java/gov/nasa/jpl/ml/cmdline/authorage/AgeClassifyTrainerTool.java
@@ -34,7 +34,7 @@ import opennlp.tools.cmdline.CmdLineUtil;
 import opennlp.tools.cmdline.ObjectStreamFactory;
 import opennlp.tools.cmdline.StreamFactoryRegistry;
 import opennlp.tools.cmdline.TerminateToolException;
-import opennlp.tools.ml.AgeClassifyTrainerFactory;
+import opennlp.tools.ml.authorage.AgeClassifyTrainerFactory;
 import opennlp.tools.tokenize.Tokenizer;
 import opennlp.tools.tokenize.WhitespaceTokenizer;
 import opennlp.tools.util.TrainingParameters;
