+++ src/main/java/gov/nasa/jpl/ml/spark/authorage/AgeClassifySparkTrainer.java
@@ -31,8 +31,8 @@ import opennlp.tools.authorage.AgeClassifyFactory;
 import opennlp.tools.authorage.AgeClassifyModel;
 import opennlp.tools.cmdline.CmdLineUtil;
 import opennlp.tools.cmdline.TerminateToolException;
-import opennlp.tools.ml.AgeClassifyTrainerFactory;
 import opennlp.tools.ml.EventTrainer;
+import opennlp.tools.ml.authorage.AgeClassifyTrainerFactory;
 import opennlp.tools.ml.model.Event;
 import opennlp.tools.ml.model.MaxentModel;
 import opennlp.tools.util.ObjectStream;
